refactor: refactor drawing related modules
