+++ packages/docs-drawing-ui/src/commands/commands/set-drawing-arrange.command.ts
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-import type { ICommand, IDrawingOrderMapParam, Nullable } from '@univerjs/core';
-import { ArrangeTypeEnum,
+import type { ICommand, Nullable } from '@univerjs/core';
+import {
     CommandType,
     ICommandService,
     IUndoRedoService,
 } from '@univerjs/core';
 import { DocDrawingApplyType, IDocDrawingService, SetDocDrawingApplyMutation } from '@univerjs/docs-drawing';
-import type { IDrawingJsonUndo1 } from '@univerjs/drawing';
+import type { IDrawingJsonUndo1, IDrawingOrderMapParam } from '@univerjs/drawing';
+import { ArrangeTypeEnum } from '@univerjs/drawing';
 import type { IAccessor } from '@wendellhu/redi';
 
 export interface ISetDrawingArrangeCommandParams extends IDrawingOrderMapParam {
