+++ packages/docs-drawing-ui/src/commands/operations/edit-doc-drawing.operation.ts
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-import type { IDrawingSearch, IOperation } from '@univerjs/core';
+import type { IOperation } from '@univerjs/core';
 import {
     CommandType,
     ICommandService,
-    IDrawingManagerService,
 
 } from '@univerjs/core';
 import type { IAccessor } from '@wendellhu/redi';
+import type { IDrawingSearch } from '@univerjs/drawing';
+import { IDrawingManagerService } from '@univerjs/drawing';
 import { SidebarDocDrawingOperation } from './open-drawing-panel.operation';
 
 export const EditDocDrawingOperation: IOperation<IDrawingSearch> = {
