+++ packages/drawing-ui/src/controllers/utils.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-import type { IDrawingManagerService, IDrawingSearch, IUniverInstanceService, Nullable, Workbook } from '@univerjs/core';
+import type { IUniverInstanceService, Nullable, Workbook } from '@univerjs/core';
 import { UniverInstanceType } from '@univerjs/core';
+import type { IDrawingManagerService, IDrawingSearch } from '@univerjs/drawing';
 import { getDrawingShapeKeyByDrawingSearch } from '@univerjs/drawing';
 import type { BaseObject, Scene } from '@univerjs/engine-render';
 import { DRAWING_OBJECT_LAYER_INDEX, Group } from '@univerjs/engine-render';
