+++ packages/drawing/src/index.ts
@@ -15,9 +15,30 @@
  */
 
 export { DRAWING_IMAGE_WIDTH_LIMIT, DRAWING_IMAGE_HEIGHT_LIMIT, DRAWING_IMAGE_COUNT_LIMIT, DRAWING_IMAGE_ALLOW_SIZE, DRAWING_IMAGE_ALLOW_IMAGE_LIST } from './basics/config';
-export { DrawingManagerService, UnitDrawingService, type IDrawingJson1Type, type IDrawingJsonUndo1 } from './services/drawing-manager.service';
+export { DrawingManagerService, UnitDrawingService, type IDrawingJson1Type, type IDrawingJsonUndo1 } from './services/drawing-manager-impl.service';
 export { getDrawingShapeKeyByDrawingSearch } from './utils/get-image-shape-key';
 export { getImageSize } from './utils/get-image-size';
 export { UniverDrawingPlugin } from './plugin';
 export { type IImageData } from './models/image-model-interface';
-
+export { IImageIoService, type IImageIoServiceParam, ImageSourceType, ImageUploadStatusType } from './services/image-io.service';
+export { ImageIoService } from './services/image-io-impl.service';
+export type { ITransformState, IRotationSkewFlipTransform, ISrcRect } from './services/interface';
+export {
+    DEFAULT_DOCUMENT_SUB_COMPONENT_ID,
+    type DrawingType,
+    DrawingTypeEnum,
+    type IDrawingSearch,
+    type IDrawingSpace,
+    type IDrawingParam,
+    IDrawingManagerService,
+    type IDrawingOrderUpdateParam,
+    type IUnitDrawingService,
+    type IDrawingMap,
+    type IDrawingSubunitMap,
+    type IDrawingOrderMapParam,
+    type IDrawingGroupUpdateParam,
+    type IDrawingVisibleParam,
+    type IDrawingMapItemData,
+    type IDrawingMapItem,
+    ArrangeTypeEnum,
+} from './services/drawing-manager.service';
