+++ packages/drawing/src/services/image-io.service.ts
@@ -14,84 +14,42 @@
  * limitations under the License.
  */
 
-import type { IImageIoService, IImageIoServiceParam, Nullable } from '@univerjs/core';
-import { ImageSourceType, ImageUploadStatusType, Tools } from '@univerjs/core';
+import type { Nullable } from '@univerjs/core';
+import { createIdentifier } from '@wendellhu/redi';
 import type { Observable } from 'rxjs';
-import { Subject } from 'rxjs';
-import { DRAWING_IMAGE_ALLOW_IMAGE_LIST, DRAWING_IMAGE_ALLOW_SIZE } from '../basics/config';
 
-export class ImageIoService implements IImageIoService {
-    private _waitCount = 0;
-
-    private _change$ = new Subject<number>();
-    change$ = this._change$ as Observable<number>;
-
-    setWaitCount(count: number) {
-        this._waitCount = count;
-        this._change$.next(count);
-    }
-
-    private _imageSourceCache: Map<string, HTMLImageElement> = new Map();
-    getImageSourceCache(source: string, imageSourceType: ImageSourceType) {
-        if (imageSourceType === ImageSourceType.BASE64) {
-            const image = new Image();
-            image.src = source;
-            return image;
-        }
-        return this._imageSourceCache.get(source);
-    }
-
-    addImageSourceCache(source: string, imageSourceType: ImageSourceType, imageSource: Nullable<HTMLImageElement>) {
-        if (imageSourceType === ImageSourceType.BASE64 || imageSource == null) {
-            return;
-        }
-        this._imageSourceCache.set(source, imageSource);
-    }
-
-    async getImage(imageId: string): Promise<string> {
-        return Promise.resolve(imageId);
-    }
-
-    async saveImage(imageFile: File): Promise<Nullable<IImageIoServiceParam>> {
-        return new Promise((resolve, reject) => {
-            if (!DRAWING_IMAGE_ALLOW_IMAGE_LIST.includes(imageFile.type)) {
-                reject(new Error(ImageUploadStatusType.ERROR_IMAGE_TYPE));
-                this._decreaseWaiting();
-                return;
-            }
-            if (imageFile.size > DRAWING_IMAGE_ALLOW_SIZE) {
-                reject(new Error(ImageUploadStatusType.ERROR_EXCEED_SIZE));
-                this._decreaseWaiting();
-                return;
-            }
-            // 获取上传的图片的宽高
-            const reader = new FileReader();
-            reader.readAsDataURL(imageFile);
-            reader.onload = (evt) => {
-                const replaceSrc = evt.target?.result as string;
-                if (replaceSrc == null) {
-                    reject(new Error(ImageUploadStatusType.ERROR_IMAGE));
-                    this._decreaseWaiting();
-                    return;
-                }
-
-                const imageId = Tools.generateRandomId(6);
-                resolve({
-                    imageId,
-                    imageSourceType: ImageSourceType.BASE64,
-                    source: replaceSrc,
-                    base64Cache: replaceSrc,
-                    status: ImageUploadStatusType.SUCCUSS,
-                });
-
-                this._decreaseWaiting();
-            };
-        });
-    }
-
-    private _decreaseWaiting() {
-        this._waitCount -= 1;
-        this._change$.next(this._waitCount);
-    }
+export enum ImageSourceType {
+    URL = 'URL',
+    UUID = 'UUID',
+    BASE64 = 'BASE64',
 }
 
+export enum ImageUploadStatusType {
+    SUCCUSS = '0',
+    ERROR_EXCEED_SIZE = '1',
+    ERROR_IMAGE_TYPE = '2',
+    ERROR_UPLOAD_COUNT_LIMIT = '3',
+    ERROR_IMAGE = '4',
+}
+
+export interface IImageIoServiceParam {
+    imageId: string;
+    imageSourceType: ImageSourceType;
+    source: string;
+    base64Cache: string;
+    status: ImageUploadStatusType;
+}
+
+export interface IImageIoService {
+    change$: Observable<number>;
+    setWaitCount(count: number): void;
+
+    getImage(imageId: string): Promise<string>;
+
+    saveImage(imageFile: File): Promise<Nullable<IImageIoServiceParam>>;
+
+    getImageSourceCache(source: string, imageSourceType: ImageSourceType): Nullable<HTMLImageElement>;
+    addImageSourceCache(source: string, imageSourceType: ImageSourceType, imageSource: Nullable<HTMLImageElement>): void;
+}
+
+export const IImageIoService = createIdentifier<IImageIoService>('core.image-io.service');
