+++ packages/sheets-drawing-ui/src/commands/operations/open-drawing-panel.operation.ts
@@ -15,10 +15,11 @@
  */
 
 import type { ICommand } from '@univerjs/core';
-import { CommandType, IDrawingManagerService, IUniverInstanceService, LocaleService } from '@univerjs/core';
+import { CommandType, IUniverInstanceService, LocaleService } from '@univerjs/core';
 import { ISidebarService } from '@univerjs/ui';
 import type { IAccessor } from '@wendellhu/redi';
 import { getSheetCommandTarget } from '@univerjs/sheets';
+import { IDrawingManagerService } from '@univerjs/drawing';
 import { COMPONENT_SHEET_DRAWING_PANEL } from '../../views/sheet-image-panel/component-name';
 
 export interface IUIComponentCommandParams {
