+++ packages/sheets-drawing-ui/src/views/sheet-image-panel/SheetDrawingAnchor.tsx
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import type { IDrawingParam, Nullable } from '@univerjs/core';
-import { ICommandService, IDrawingManagerService, LocaleService } from '@univerjs/core';
+import type { Nullable } from '@univerjs/core';
+import { ICommandService, LocaleService } from '@univerjs/core';
 import { useDependency } from '@wendellhu/redi/react-bindings';
 import React, { useEffect, useState } from 'react';
 import { Radio, RadioGroup } from '@univerjs/design';
@@ -24,6 +24,7 @@ import type { ISheetDrawing } from '@univerjs/sheets-drawing';
 import { SheetDrawingAnchorType } from '@univerjs/sheets-drawing';
 import type { BaseObject } from '@univerjs/engine-render';
 import { IRenderManagerService } from '@univerjs/engine-render';
+import { IDrawingManagerService, type IDrawingParam } from '@univerjs/drawing';
 import { SetSheetDrawingCommand } from '../../commands/commands/set-sheet-drawing.command';
 import styles from './index.module.less';
