+++ packages/sheets-drawing/src/commands/mutations/set-drawing-apply.mutation.ts
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-import type { IDrawingGroupUpdateParam, IDrawingOrderMapParam, IDrawingSearch, IMutation } from '@univerjs/core';
-import { CommandType, IDrawingManagerService } from '@univerjs/core';
-import type { IDrawingJson1Type } from '@univerjs/drawing';
+import type { IDrawingSearch, IMutation } from '@univerjs/core';
+import { CommandType } from '@univerjs/core';
+import type { IDrawingGroupUpdateParam, IDrawingJson1Type, IDrawingOrderMapParam } from '@univerjs/drawing';
+import { IDrawingManagerService } from '@univerjs/drawing';
 import { ISheetDrawingService } from '../../services/sheet-drawing.service';
 
 export enum DrawingApplyType {
