+++ packages/sheets-drawing/src/controllers/sheet-drawing.controller.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-import type { IDrawingSubunitMap } from '@univerjs/core';
-import { Disposable, ICommandService, IDrawingManagerService, IResourceManagerService, LifecycleStages, OnLifecycle, UniverInstanceType } from '@univerjs/core';
+import { Disposable, ICommandService, IResourceManagerService, LifecycleStages, OnLifecycle, UniverInstanceType } from '@univerjs/core';
+import type { IDrawingSubunitMap } from '@univerjs/drawing';
+import { IDrawingManagerService } from '@univerjs/drawing';
 import type { ISheetDrawing } from '../services/sheet-drawing.service';
 import { ISheetDrawingService } from '../services/sheet-drawing.service';
 import { SetDrawingApplyMutation } from '../commands/mutations/set-drawing-apply.mutation';
